requirements.txt
alembic==1.8.1
aniso8601==9.0.1
attrs==22.1.0
click==8.1.3
Flask==2.2.2
Flask-JWT-Extended==4.4.4
Flask-Migrate==4.0.0
flask-restx==1.0.3
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.0
Mako==1.2.4
MarkupSafe==2.1.1
PyJWT==2.6.0
pyrsistent==0.19.2
python-decouple==3.6
pytz==2022.6
SQLAlchemy==1.4.44
Werkzeug==2.2.2
zipp==3.10.0
